requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
SQLAlchemy==2.0.35
Flask-SQLAlchemy==3.1.1
alembic==1.13.2
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
cryptography==43.0.1
email-validator==2.2.0
